Revert "Add independent Go modules for terragrunt and dependencies"
This reverts PR #1608, which introduced independent Go modules for the
terragrunt dependency chain. Reverting back to the original monolithic
module structure.
