+++ go.mod
@@ -1,6 +1,8 @@
 module github.com/gruntwork-io/terratest
 
-go 1.25.0
+go 1.24.0
+
+toolchain go1.24.2
 
 require (
 	cloud.google.com/go v0.116.0 // indirect
@@ -11,30 +13,30 @@ require (
 	github.com/Azure/go-autorest/autorest/to v0.4.0 // indirect
 	github.com/Azure/go-autorest/autorest/validation v0.3.1 // indirect
 	github.com/aws/aws-lambda-go v1.47.0
-	github.com/go-errors/errors v1.4.2 // indirect
+	github.com/go-errors/errors v1.0.2-0.20180813162953-d98b870cc4e0 // indirect
 	github.com/go-sql-driver/mysql v1.8.1
 	github.com/google/go-containerregistry v0.20.2
 	github.com/google/uuid v1.6.0
 	github.com/gruntwork-io/go-commons v0.8.0
 	github.com/hashicorp/go-multierror v1.1.1
 	github.com/hashicorp/go-version v1.7.0
-	github.com/hashicorp/hcl/v2 v2.22.0 // indirect
-	github.com/hashicorp/terraform-json v0.23.0 // indirect
-	github.com/jinzhu/copier v0.0.0-20190924061706-b57f9002281a // indirect
-	github.com/jstemmer/go-junit-report v1.0.0 // indirect
+	github.com/hashicorp/hcl/v2 v2.22.0
+	github.com/hashicorp/terraform-json v0.23.0
+	github.com/jinzhu/copier v0.0.0-20190924061706-b57f9002281a
+	github.com/jstemmer/go-junit-report v1.0.0
 	github.com/magiconair/properties v1.8.7
-	github.com/mattn/go-zglob v0.0.6
+	github.com/mattn/go-zglob v0.0.2-0.20190814121620-e3c945676326
 	github.com/miekg/dns v1.1.62
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/oracle/oci-go-sdk v7.1.0+incompatible
 	github.com/pquerna/otp v1.4.0
 	github.com/sirupsen/logrus v1.9.3
-	github.com/stretchr/testify v1.11.1
-	github.com/tmccombs/hcl2json v0.6.4 // indirect
+	github.com/stretchr/testify v1.10.0
+	github.com/tmccombs/hcl2json v0.6.4
 	github.com/urfave/cli v1.22.16
-	github.com/zclconf/go-cty v1.15.0 // indirect
-	golang.org/x/crypto v0.43.0
-	golang.org/x/net v0.46.0 // indirect
+	github.com/zclconf/go-cty v1.15.0
+	golang.org/x/crypto v0.41.0
+	golang.org/x/net v0.43.0
 	golang.org/x/oauth2 v0.27.0
 	google.golang.org/api v0.206.0
 	google.golang.org/genproto v0.0.0-20241113202542-65e8d215514f
@@ -73,12 +75,6 @@ require (
 	github.com/aws/aws-sdk-go-v2/service/sts v1.33.1
 	github.com/denisenkom/go-mssqldb v0.12.3
 	github.com/gonvenience/ytbx v1.4.4
-	github.com/gruntwork-io/terratest/modules/files v0.0.0
-	github.com/gruntwork-io/terratest/modules/logger v0.0.0
-	github.com/gruntwork-io/terratest/modules/retry v0.0.0
-	github.com/gruntwork-io/terratest/modules/shell v0.0.0
-	github.com/gruntwork-io/terratest/modules/terraform v0.0.0-00010101000000-000000000000
-	github.com/gruntwork-io/terratest/modules/testing v0.0.0
 	github.com/hashicorp/go-getter/v2 v2.2.3
 	github.com/homeport/dyff v1.6.0
 	github.com/jackc/pgx/v5 v5.7.1
@@ -164,7 +160,6 @@ require (
 	github.com/googleapis/enterprise-certificate-proxy v0.3.4 // indirect
 	github.com/googleapis/gax-go/v2 v2.14.0 // indirect
 	github.com/gorilla/websocket v1.5.4-0.20250319132907-e064f32e3674 // indirect
-	github.com/gruntwork-io/terratest/modules/formatting v0.0.0 // indirect
 	github.com/hashicorp/errwrap v1.0.0 // indirect
 	github.com/hashicorp/go-cleanhttp v0.5.2 // indirect
 	github.com/hashicorp/go-safetemp v1.0.0 // indirect
@@ -174,7 +169,7 @@ require (
 	github.com/jmespath/go-jmespath v0.4.0 // indirect
 	github.com/josharian/intern v1.0.0 // indirect
 	github.com/json-iterator/go v1.1.12 // indirect
-	github.com/klauspost/compress v1.16.7 // indirect
+	github.com/klauspost/compress v1.16.5 // indirect
 	github.com/kylelemons/godebug v1.1.0 // indirect
 	github.com/lucasb-eyer/go-colorful v1.2.0 // indirect
 	github.com/mailru/easyjson v0.7.7 // indirect
@@ -215,13 +210,13 @@ require (
 	go.opentelemetry.io/otel/trace v1.29.0 // indirect
 	go.yaml.in/yaml/v2 v2.4.2 // indirect
 	go.yaml.in/yaml/v3 v3.0.4 // indirect
-	golang.org/x/mod v0.28.0 // indirect
-	golang.org/x/sync v0.17.0 // indirect
-	golang.org/x/sys v0.37.0 // indirect
-	golang.org/x/term v0.36.0 // indirect
-	golang.org/x/text v0.30.0 // indirect
+	golang.org/x/mod v0.27.0 // indirect
+	golang.org/x/sync v0.16.0 // indirect
+	golang.org/x/sys v0.35.0 // indirect
+	golang.org/x/term v0.34.0 // indirect
+	golang.org/x/text v0.28.0 // indirect
 	golang.org/x/time v0.9.0 // indirect
-	golang.org/x/tools v0.37.0 // indirect
+	golang.org/x/tools v0.36.0 // indirect
 	google.golang.org/genproto/googleapis/api v0.0.0-20241104194629-dd2ea8efbc28 // indirect
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20241104194629-dd2ea8efbc28 // indirect
 	google.golang.org/grpc v1.67.1 // indirect
@@ -237,15 +232,4 @@ require (
 	sigs.k8s.io/randfill v1.0.0 // indirect
 	sigs.k8s.io/structured-merge-diff/v6 v6.3.0 // indirect
 	sigs.k8s.io/yaml v1.6.0 // indirect
-)
-
-replace (
-	github.com/gruntwork-io/terratest/modules/files => ./modules/files
-	github.com/gruntwork-io/terratest/modules/formatting => ./modules/formatting
-	github.com/gruntwork-io/terratest/modules/logger => ./modules/logger
-	github.com/gruntwork-io/terratest/modules/retry => ./modules/retry
-	github.com/gruntwork-io/terratest/modules/shell => ./modules/shell
-	github.com/gruntwork-io/terratest/modules/terraform => ./modules/terraform
-	github.com/gruntwork-io/terratest/modules/terragrunt => ./modules/terragrunt
-	github.com/gruntwork-io/terratest/modules/testing => ./modules/testing
 )
