--- modules/files/go.mod
@@ -1,17 +1,0 @@
-module github.com/gruntwork-io/terratest/modules/files
-
-go 1.25.0
-
-require (
-	github.com/mattn/go-zglob v0.0.6
-	github.com/stretchr/testify v1.11.1
-)
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/kr/pretty v0.3.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/rogpeppe/go-internal v1.13.1 // indirect
-	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
