+++ modules/terraform/format_test.go
@@ -75,6 +75,51 @@ func TestFormatTerraformVarsAsArgs(t *testing.T) {
 	}
 }
 
+func TestPrimitiveToHclString(t *testing.T) {
+	t.Parallel()
+
+	testCases := []struct {
+		value    interface{}
+		expected string
+	}{
+		{"", ""},
+		{"foo", "foo"},
+		{"true", "true"},
+		{true, "true"},
+		{3, "3"},
+		{nil, "null"},
+		{[]int{1, 2, 3}, "[1 2 3]"}, // Anything that isn't a primitive is forced into a string
+	}
+
+	for _, testCase := range testCases {
+		actual := primitiveToHclString(testCase.value, false)
+		assert.Equal(t, testCase.expected, actual, "Value: %v", testCase.value)
+	}
+}
+
+func TestMapToHclString(t *testing.T) {
+	t.Parallel()
+
+	testCases := []struct {
+		value    map[string]interface{}
+		expected string
+	}{
+		{map[string]interface{}{}, "{}"},
+		{map[string]interface{}{"key1": "value1"}, "{\"key1\" = \"value1\"}"},
+		{map[string]interface{}{"key1": 123}, "{\"key1\" = 123}"},
+		{map[string]interface{}{"key1": true}, "{\"key1\" = true}"},
+		{map[string]interface{}{"key1": []int{1, 2, 3}}, "{\"key1\" = [1, 2, 3]}"}, // Any value that isn't a primitive is forced into a string
+		{map[string]interface{}{"key1": "value1", "key2": 0, "key3": false}, "{\"key1\" = \"value1\", \"key2\" = 0, \"key3\" = false}"},
+		{map[string]interface{}{"key1.a.b.c": "value1"}, "{\"key1.a.b.c\" = \"value1\"}"},
+	}
+
+	for _, testCase := range testCases {
+		checkResultWithRetry(t, 100, testCase.expected, fmt.Sprintf("mapToHclString(%v)", testCase.value), func() interface{} {
+			return mapToHclString(testCase.value)
+		})
+	}
+}
+
 // Some of our tests execute code that loops over a map to produce output. The problem is that the order of map
 // iteration is generally unpredictable and, to make it even more unpredictable, Go intentionally randomizes the
 // iteration order (https://blog.golang.org/go-maps-in-action#TOC_7). Therefore, the order of items in the output
@@ -105,6 +150,114 @@ func checkResultWithRetry(t *testing.T, maxRetries int, expectedValue interface{
 	}
 
 	assert.Fail(t, "checkResultWithRetry failed", "After %d retries, %s still not succeeding (see retries above)", description)
+}
+
+func TestSliceToHclString(t *testing.T) {
+	t.Parallel()
+
+	testCases := []struct {
+		value    []interface{}
+		expected string
+	}{
+		{[]interface{}{}, "[]"},
+		{[]interface{}{"foo"}, "[\"foo\"]"},
+		{[]interface{}{123}, "[123]"},
+		{[]interface{}{true}, "[true]"},
+		{[]interface{}{[]int{1, 2, 3}}, "[[1, 2, 3]]"}, // Any value that isn't a primitive is forced into a string
+		{[]interface{}{"foo", 0, false}, "[\"foo\", 0, false]"},
+		{[]interface{}{map[string]interface{}{"foo": "bar"}}, "[{\"foo\" = \"bar\"}]"},
+		{[]interface{}{map[string]interface{}{"foo": "bar"}, map[string]interface{}{"foo": "bar"}}, "[{\"foo\" = \"bar\"}, {\"foo\" = \"bar\"}]"},
+	}
+
+	for _, testCase := range testCases {
+		actual := sliceToHclString(testCase.value)
+		assert.Equal(t, testCase.expected, actual, "Value: %v", testCase.value)
+	}
+}
+
+func TestToHclString(t *testing.T) {
+	t.Parallel()
+
+	testCases := []struct {
+		value    interface{}
+		expected string
+	}{
+		{"", ""},
+		{"foo", "foo"},
+		{123, "123"},
+		{true, "true"},
+		{[]int{1, 2, 3}, "[1, 2, 3]"},
+		{[]string{"foo", "bar", "baz"}, "[\"foo\", \"bar\", \"baz\"]"},
+		{map[string]string{"key1": "value1"}, "{\"key1\" = \"value1\"}"},
+		{map[string]int{"key1": 123}, "{\"key1\" = 123}"},
+	}
+
+	for _, testCase := range testCases {
+		actual := toHclString(testCase.value, false)
+		assert.Equal(t, testCase.expected, actual, "Value: %v", testCase.value)
+	}
+}
+
+func TestTryToConvertToGenericSlice(t *testing.T) {
+	t.Parallel()
+
+	testCases := []struct {
+		value           interface{}
+		expectedSlice   []interface{}
+		expectedIsSlice bool
+	}{
+		{"", []interface{}{}, false},
+		{"foo", []interface{}{}, false},
+		{true, []interface{}{}, false},
+		{531, []interface{}{}, false},
+		{map[string]string{"foo": "bar"}, []interface{}{}, false},
+		{[]string{}, []interface{}{}, true},
+		{[]int{}, []interface{}{}, true},
+		{[]bool{}, []interface{}{}, true},
+		{[]interface{}{}, []interface{}{}, true},
+		{[]string{"foo", "bar", "baz"}, []interface{}{"foo", "bar", "baz"}, true},
+		{[]int{1, 2, 3}, []interface{}{1, 2, 3}, true},
+		{[]bool{true, true, false}, []interface{}{true, true, false}, true},
+		{[]interface{}{"foo", "bar", "baz"}, []interface{}{"foo", "bar", "baz"}, true},
+	}
+
+	for _, testCase := range testCases {
+		actualSlice, actualIsSlice := tryToConvertToGenericSlice(testCase.value)
+		assert.Equal(t, testCase.expectedSlice, actualSlice, "Value: %v", testCase.value)
+		assert.Equal(t, testCase.expectedIsSlice, actualIsSlice, "Value: %v", testCase.value)
+	}
+}
+
+func TestTryToConvertToGenericMap(t *testing.T) {
+	t.Parallel()
+
+	testCases := []struct {
+		value         interface{}
+		expectedMap   map[string]interface{}
+		expectedIsMap bool
+	}{
+		{"", map[string]interface{}{}, false},
+		{"foo", map[string]interface{}{}, false},
+		{true, map[string]interface{}{}, false},
+		{531, map[string]interface{}{}, false},
+		{[]string{"foo", "bar"}, map[string]interface{}{}, false},
+		{map[int]int{}, map[string]interface{}{}, false},
+		{map[bool]string{}, map[string]interface{}{}, false},
+		{map[string]string{}, map[string]interface{}{}, true},
+		{map[string]int{}, map[string]interface{}{}, true},
+		{map[string]bool{}, map[string]interface{}{}, true},
+		{map[string]interface{}{}, map[string]interface{}{}, true},
+		{map[string]string{"key1": "value1", "key2": "value2"}, map[string]interface{}{"key1": "value1", "key2": "value2"}, true},
+		{map[string]int{"key1": 1, "key2": 2, "key3": 3}, map[string]interface{}{"key1": 1, "key2": 2, "key3": 3}, true},
+		{map[string]bool{"key1": true}, map[string]interface{}{"key1": true}, true},
+		{map[string]interface{}{"key1": "value1"}, map[string]interface{}{"key1": "value1"}, true},
+	}
+
+	for _, testCase := range testCases {
+		actualMap, actualIsMap := tryToConvertToGenericMap(testCase.value)
+		assert.Equal(t, testCase.expectedMap, actualMap, "Value: %v", testCase.value)
+		assert.Equal(t, testCase.expectedIsMap, actualIsMap, "Value: %v", testCase.value)
+	}
 }
 
 func TestFormatArgsAppliesLockCorrectly(t *testing.T) {
