--- modules/testing/go.mod
@@ -1,3 +1,0 @@
-module github.com/gruntwork-io/terratest/modules/testing
-
-go 1.25.0
